Created *node_type and simplified node creation
node_type allows for creating nodes whose data is other nodes.
Creating a nested node uses nody_copy which is shallow - only the data is copied, not the owner, or children.
However, if the data of a node you are copying is also a node, it will be recursively shallow copied as well.

Removed owner and table size input requirements in node_new.
Putting a node into an owner's table is the only reliable way of creating an owner/child relationship.
Having an initial table size provided by the user doesn't seem to be very useful and only complicates the function call.

Added documentation comments to node.c and node.h

Updated the root readme.