Created nested_node_type
We are no longer copying node data when creating nested nodes.
Now we simply point to an existing node that is provided and tell
the nested node whether this data should be freed along with itself
or not. This way we avoid needless copying, but can still go on
to build reusable datastructures.

The 'freev' function provided by the node's type structure must
free all data that was allocated by the type's 'new' function.
This includes freeing the 'data' pointer itself as in:

static void _free_my_node_data(void *d)
{
	_free_my_node_data_internals((struct my_node_s *) d);

	// Important!
	free(d);
}

Otherwise, we will have memory leaks. As a rule of thumb remember,
if you allocated it, it's your responsibility to free it.