Added stack and queue functionality
The stack is implemented as a circular-linked list to allow it to work as a
queue as well. The owner of the 'stack' node is the 'last' element of the
stack.

See stack.h for a description of the similarities and differences between
stack and queue functions.

Added recurse flag to node_free.

Added pr_dbg functionality.